Exception handling

+++ dscommerce/src/main/java/com/coelhodev/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.coelhodev.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.coelhodev.dscommerce.dto.ProductDTO;
 import com.coelhodev.dscommerce.entities.Product;
 import com.coelhodev.dscommerce.repositories.ProductRepository;
+import com.coelhodev.dscommerce.services.exceptions.DatabaseException;
+import com.coelhodev.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 
@@ -19,7 +25,7 @@ public class ProductService {
 	
 	@Transactional(readOnly = true)
 	public ProductDTO findById(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return new ProductDTO(product);
 	
 	}
@@ -44,17 +50,29 @@ public class ProductService {
 	
 	@Transactional
 	public ProductDTO update (Long id, ProductDTO dto) {
-		
+	
+	try {	
 		Product entity = repository.getReferenceById(id);
 		copyDtoToEntity(dto, entity);
-		
 		entity = repository.save(entity);
 		return new ProductDTO(entity);
 	}
+	catch (EntityNotFoundException e) {
+		throw new ResourceNotFoundException("Recurso não encontrado");
+	}
+	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if (!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
+	        	repository.deleteById(id);    		
+		}
+	    	catch (DataIntegrityViolationException e) {
+	        	throw new DatabaseException("Falha de integridade referencial");
+	   	}
 	}
